feature: 회원/매물/계약 본인인증 기능 연동

+++ back-end/src/main/java/com/bangjwo/contract/domain/repository/ContractRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.bangjwo.contract.domain.entity.Contract;
 
 @Repository
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 	Optional<Contract> findByContractId(Long contractId);
 }
